test: add more tests for `apply` changes, split `apply` tests in its own file
