+++ tests/test_data/imap_mag_four-vector-offsets-layer-meta_20251017_v001.json
