+++ tests/test_data/imap_mag_l1c_norm-mago-four-vectors-four-ranges_20251017_v000.cdf
