+++ tests/test_data/imap_mag_misaligned-timestamps-layer-meta_20251017_v001.json
