+++ tests/test_data/metadata_file_with_values.json
@@ -1,0 +1,199 @@
+{
+    "id": "",
+    "metadata": {
+        "comment": "This is a zero value calibration file",
+        "creation_timestamp": "2025-03-28T11:22:20",
+        "data_filename": null,
+        "dependencies": [],
+        "science": [
+            ".work/imap_mag_l1b_norm-mago_20251017_v002.json"
+        ]
+    },
+    "method": "noop",
+    "mission": "IMAP",
+    "sensor": "MAGo",
+    "validity": {
+        "end": "2025-10-17T02:11:59.021309",
+        "start": "2025-10-17T02:11:51.521"
+    },
+    "value_type": "vector",
+    "values": [
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:51.521309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:52.021309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:52.521309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:53.021309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:53.521309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:54.021309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:54.521309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:55.021309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:55.521309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:56.021309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:56.521309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:57.021309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:57.521309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:58.021309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:58.521309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        },
+        {
+            "quality_bitmask": 0.0,
+            "quality_flag": 0.0,
+            "time": "2025-10-17T02:11:59.021309",
+            "timedelta": 0.0,
+            "value": [
+                0.0,
+                0.0,
+                0.0
+            ]
+        }
+    ],
+    "version": 0.0
+}
